Move aws-sdk-1.11 library internals into .internal package; remove advice bridge
Mirrors the aws-sdk-2.2 layout: only the public `AwsSdkTelemetry` and `AwsSdkTelemetryBuilder` entry points remain at the top of the library package; the 26 supporting helpers move to the `.internal` subpackage. The classes that need to be reachable from the entry points or from the javaagent are promoted to `public` with the standard internal-use javadoc; the rest stay package-private. The `SqsAdviceBridge` indirection is replaced with a direct `SqsImpl.class.getName()` reference in `SqsInstrumentationModule.RegisterAdvice`, removing the cross-namespace javaagent source that previously lived under `io.opentelemetry.instrumentation.awssdk.v1_11`.
