--- instrumentation/aws-sdk/aws-sdk-1.11/javaagent/src/main/java/io/opentelemetry/instrumentation/awssdk/v1_11/SqsAdviceBridge.java
@@ -1,15 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.awssdk.v1_11;
-
-public class SqsAdviceBridge {
-  private SqsAdviceBridge() {}
-
-  public static void referenceForMuzzleOnly() {
-    throw new UnsupportedOperationException(
-        SqsImpl.class.getName() + " referencing for muzzle, should never be actually called");
-  }
-}
